Add files via upload
Because of importance of the notion of universal property, product and coproduct, it looks reasonable to workout problems from the Chapter II § 3 in more details by means of commutative diagrams, because diagrams help to see visually whats going on. That is up to you to decide whether include this solution into the file or not.